+++ packages/sdk/tests/e2e/dataProtectorCore/protectData.test.ts
@@ -2,12 +2,15 @@ import fsPromises from 'fs/promises';
 import path from 'path';
 import { beforeEach, describe, expect, it } from '@jest/globals';
 import { HDNodeWallet, Wallet } from 'ethers';
+import { IExec } from 'iexec';
 import { SmsCallError } from 'iexec/errors';
 import { IExecDataProtectorCore, WorkflowError } from '../../../src/index.js';
 import {
   MAX_EXPECTED_BLOCKTIME,
   MAX_EXPECTED_WEB2_SERVICES_TIME,
   getTestConfig,
+  getTestIExecOption,
+  getTestRpcProvider,
   getTestWeb3SignerProvider,
 } from '../../test-utils.js';
 
@@ -25,7 +28,7 @@ describe('dataProtectorCore.protectData()', () => {
     'creates the protected data',
     async () => {
       // load some binary data
-      await fsPromises.readFile(
+      const pngImage = await fsPromises.readFile(
         path.join(process.cwd(), 'tests', '_test_inputs_', 'image.png')
       );
       const data = {
@@ -42,8 +45,7 @@ describe('dataProtectorCore.protectData()', () => {
             with: {
               binary: {
                 data: {
-                  pngImage: 'placeholder',
-                  // pngImage, // commented as currently too large for IPFS upload
+                  pngImage,
                 },
               },
             },
@@ -67,8 +69,7 @@ describe('dataProtectorCore.protectData()', () => {
             with: {
               binary: {
                 data: {
-                  pngImage: 'string',
-                  // pngImage: '', // commented as currently too large for IPFS upload
+                  pngImage: 'image/png',
                 },
               },
             },
@@ -88,6 +89,21 @@ describe('dataProtectorCore.protectData()', () => {
       expect(typeof result.transactionHash).toBe('string');
       expect(result.zipFile).toBeInstanceOf(Uint8Array);
       expect(typeof result.encryptionKey).toBe('string');
+
+      const ethProvider = getTestRpcProvider();
+      const iexecOptions = getTestIExecOption();
+      const iexecProd = new IExec({ ethProvider }, iexecOptions);
+      const iexecDebug = new IExec(
+        { ethProvider },
+        { ...iexecOptions, smsURL: iexecOptions.smsDebugURL }
+      );
+      const prodSecretPushed = await iexecProd.dataset.checkDatasetSecretExists(
+        result.address
+      );
+      const debugSecretPushed =
+        await iexecDebug.dataset.checkDatasetSecretExists(result.address);
+      expect(prodSecretPushed).toBe(true);
+      expect(debugSecretPushed).toBe(false);
     },
     2 * MAX_EXPECTED_BLOCKTIME + MAX_EXPECTED_WEB2_SERVICES_TIME
   );
@@ -99,6 +115,7 @@ describe('dataProtectorCore.protectData()', () => {
         getTestWeb3SignerProvider(wallet.privateKey),
         {
           iexecOptions: {
+            ...getTestIExecOption(),
             smsURL: 'https://unavailable.sms.url',
           },
         }
@@ -125,4 +142,80 @@ describe('dataProtectorCore.protectData()', () => {
     },
     2 * MAX_EXPECTED_BLOCKTIME + MAX_EXPECTED_WEB2_SERVICES_TIME
   );
+
+  describe('when `allowDebug: true`', () => {
+    it(
+      'creates the protected data with secret pushed on both prod and debug SMS',
+      async () => {
+        const data = {
+          string: 'hello world!',
+        };
+        const DATA_NAME = 'test do not use';
+
+        const result = await dataProtectorCore.protectData({
+          data,
+          name: DATA_NAME,
+          allowDebug: true,
+        });
+        expect(result.name).toBe(DATA_NAME);
+        expect(typeof result.address).toBe('string');
+        expect(result.owner).toBe(wallet.address);
+        expect(typeof result.creationTimestamp).toBe('number');
+        expect(typeof result.transactionHash).toBe('string');
+        expect(result.zipFile).toBeInstanceOf(Uint8Array);
+        expect(typeof result.encryptionKey).toBe('string');
+
+        const ethProvider = getTestRpcProvider();
+        const iexecOptions = getTestIExecOption();
+        const iexecProd = new IExec({ ethProvider }, iexecOptions);
+        const iexecDebug = new IExec(
+          { ethProvider },
+          { ...iexecOptions, smsURL: iexecOptions.smsDebugURL }
+        );
+        const prodSecretPushed =
+          await iexecProd.dataset.checkDatasetSecretExists(result.address);
+        const debugSecretPushed =
+          await iexecDebug.dataset.checkDatasetSecretExists(result.address);
+        expect(prodSecretPushed).toBe(true);
+        expect(debugSecretPushed).toBe(true);
+      },
+      2 * MAX_EXPECTED_BLOCKTIME + MAX_EXPECTED_WEB2_SERVICES_TIME
+    );
+
+    it(
+      'should throw error when debug sms is not available',
+      async () => {
+        const unavailableDataProtector = new IExecDataProtectorCore(
+          getTestWeb3SignerProvider(wallet.privateKey),
+          {
+            iexecOptions: {
+              ...getTestIExecOption(),
+              smsDebugURL: 'https://unavailable.debug-sms.url',
+            },
+          }
+        );
+        let error: WorkflowError | undefined;
+        try {
+          await unavailableDataProtector.protectData({
+            data: { doNotUse: 'test' },
+            allowDebug: true,
+          });
+        } catch (e) {
+          error = e as WorkflowError;
+        }
+        expect(error).toBeInstanceOf(WorkflowError);
+        expect(error.message).toBe(
+          "A service in the iExec protocol appears to be unavailable. You can retry later or contact iExec's technical support for help."
+        );
+        expect(error.cause).toStrictEqual(
+          new SmsCallError(
+            'Connection to https://unavailable.debug-sms.url failed with a network error',
+            Error('')
+          )
+        );
+        expect(error.isProtocolError).toBe(true);
+      },
+      2 * MAX_EXPECTED_BLOCKTIME + MAX_EXPECTED_WEB2_SERVICES_TIME
+    );
+  });
 });